requirements.txt
Django==4.2.24
djangorestframework==3.15.2
djangorestframework-simplejwt==4.0.0
razorpay